+++ theme/template/functions/index.php
@@ -11,4 +11,8 @@ include_once dirname(__FILE__) . '/blocks.php';
 include_once dirname(__FILE__) . '/categories.php';
 include_once dirname(__FILE__) . '/features.php';
 include_once dirname(__FILE__) . '/post-session.php';
+include_once dirname(__FILE__) . '/post-supporter-gold.php';
+include_once dirname(__FILE__) . '/post-supporter-operation.php';
+include_once dirname(__FILE__) . '/post-supporter-platinum.php';
+include_once dirname(__FILE__) . '/post-supporter-silver.php';
 include_once dirname(__FILE__) . '/social.php';