+++ theme/CHANGES.md
@@ -4,6 +4,7 @@
 
 ## 0.13.2
 
+- Feature: Created custom post for each supporter types
 - Refactor: Renamed sponsor to supporter
 
 ## 0.13.1